Move Rulebook-related classes to a separate package
